requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.5
djangorestframework==3.14.0
Faker==15.3.4
pycodestyle==2.10.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
